move stripThinkBlock to format.go
